requirements.txt
opencv-python>=4.8
numpy>=1.24
onnxruntime-gpu>=1.16
torch>=2.0
scipy>=1.11
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.5
pyyaml>=6.0
python-multipart>=0.0.6
ultralytics>=8.4
requests>=2.28
